+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][unreleased]
 
+## [0.4.1] - 2022-12-06
 ### Fixed
 - Better error message for `DisableMethodCallInContextRule`
 
@@ -33,7 +34,8 @@
 - Rule for checking trait context 
 - Phpstorm meta dynamic return types
 
-[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.0...HEAD
+[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.1...HEAD
+[0.4.1]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.1.0...0.2.0